+++ src/org/zaproxy/zap/extension/alert/ExtensionAlert.java
@@ -186,7 +186,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
 
             alert.setSourceHistoryId(ref.getHistoryId());
 
-            hrefs.put(Integer.valueOf(ref.getHistoryId()), ref);
+            hrefs.put(ref.getHistoryId(), ref);
             
             this.applyOverrides(alert);
 
@@ -479,7 +479,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
         final ExtensionHistory extensionHistory = Control.getSingleton().getExtensionLoader().getExtension(ExtensionHistory.class);
 
         for (int i = 0; i < v.size(); i++) {
-            int alertId = v.get(i).intValue();
+            int alertId = v.get(i);
             RecordAlert recAlert = tableAlert.read(alertId);
             int historyId = recAlert.getHistoryId();
             HistoryReference historyReference = null;
@@ -488,7 +488,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
             }
 
             if (historyReference == null) {
-                historyReference = this.hrefs.get(Integer.valueOf(historyId));
+                historyReference = this.hrefs.get(historyId);
             }
 
             Alert alert;
@@ -501,7 +501,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
             if (historyReference != null) {
                 // The ref can be null if hrefs are purged
                 addAlertToTree(alert);
-                Integer key = Integer.valueOf(historyId);
+                Integer key = historyId;
                 if (!hrefs.containsKey(key)) {
                     this.hrefs.put(key, alert.getHistoryRef());
                 }
@@ -620,7 +620,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
                 }
             }
             for (HistoryReference href : toDelete) {
-                hrefs.remove(Integer.valueOf(href.getHistoryId()));
+                hrefs.remove(href.getHistoryId());
             }
         }
 
@@ -657,7 +657,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
             this.recalcAlerts();
         }
 
-        hrefs.remove(Integer.valueOf(hRef.getHistoryId()));
+        hrefs.remove(hRef.getHistoryId());
     }
     
     /**
@@ -714,7 +714,7 @@ public class ExtensionAlert extends ExtensionAdaptor implements
             v = tableAlert.getAlertList();
 
             for (int i = 0; i < v.size(); i++) {
-                int alertId = v.get(i).intValue();
+                int alertId = v.get(i);
                 RecordAlert recAlert = tableAlert.read(alertId);
                 Alert alert = new Alert(recAlert);
                 if (alert.getHistoryRef() != null) {
